added support for gimbal android sdk and updated gimbal ios sdk
