+++ plugin.xml
@@ -2,7 +2,7 @@
 <plugin xmlns="http://www.phonegap.com/ns/plugins/1.0"
     xmlns:android="http://schemas.android.com/apk/res/android"
     id="urbanairship-gimbal-sdk-cordova"
-    version="1.0.0">
+    version="2.0.0">
 
     <name>Gimbal Plugin</name>
     <description>Cordova plugin to install the Gimbal SDK</description>
@@ -29,5 +29,50 @@
         <!-- system frameworks -->
         <framework src="CoreLocation.framework" />
         <framework src="libz.dylib" />
+    </platform>
+    
+    <!-- android -->
+    <platform name="android">
+		<config-file target="res/xml/config.xml" parent="/*">
+			<feature name="Gimbal">
+				<param name="android-package" value="com.gimbal.android.Gimbal" />
+				<param name="onload" value="true" />
+			</feature>
+		</config-file>
+		
+		<config-file target="AndroidManifest.xml" parent="/manifest">
+			<uses-permission android:name="android.permission.INTERNET" />
+			<uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+			<uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
+			<uses-permission android:name="android.permission.WAKE_LOCK" />
+			<uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+			<uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
+			<uses-permission android:name="android.permission.BLUETOOTH" />
+			<uses-permission android:name="android.permission.BLUETOOTH_ADMIN" />
+		
+			<service
+				android:name="com.gimbal.internal.service.GimbalService"
+				android:exported="false" >
+				<intent-filter>
+					<action android:name="$PACKAGE_NAME.service.GIMBAL_SERVICE" />
+				</intent-filter>
+			</service>
+			
+			<receiver
+				android:name="com.gimbal.internal.service.GimbalServiceStartStopReceiver"
+				android:enabled="true" >
+				<intent-filter>
+					<action android:name="android.intent.action.BOOT_COMPLETED" />
+				</intent-filter>
+				<intent-filter>
+					<action android:name="android.intent.action.ACTION_SHUTDOWN" />
+				</intent-filter>
+			</receiver>
+		</config-file>
+		
+		<!-- Gimbal framework -->
+		<source-file src="src/android/Gimbal.framework/Versions/2.48.1/libs/gimbal.jar" target-dir="libs"/>
+		<source-file src="src/android/Gimbal.framework/Versions/2.48.1/libs/gimbal-slf4j-impl.jar" target-dir="libs"/>
+		<source-file src="src/android/Gimbal.framework/Versions/2.48.1/libs/slf4j-api.jar" target-dir="libs"/>
     </platform>
 </plugin>
