+++ src/ios/Gimbal.framework/Versions/Current
@@ -1,1 +1,1 @@
-2.27.3
+2.48.2
